+++ src/limits.h
@@ -6,15 +6,17 @@
 #define SCALE_LOG 2
 #define SCALE_LOG_ZERO 3
 
+#include "types.h"
+
 typedef struct {
-	long long int lower;
-	long long int upper;
+	int64_t lower;
+	int64_t upper;
 } window_t;
 
 typedef struct {
-	double lower;
-	int bins;
-	double upper;
+	float64_t lower;
+	int32_t bins;
+	float64_t upper;
 } limits_t;
 
 int str_to_limits(char *str, limits_t *limits);
